# docs/swagger.yaml
openapi: 3.0.3
info:
  title: API de Gestión Documental
  version: 1.0.0
  description: |
    API para gestión documental con capacidad para gestionar subidas individuales y por colecciones a servicio externo
servers:
  - url: https://gd.tanner.qa.msys.cl/adf-tanner-service
    description: Entorno Tanner - Mycrosystem
  - url: http://localhost:5000
    description: Entorno local para desarrollo y testing
...
paths:
...
  /documentos/buscar-descargar:
    post:
      tags: [Documentos]
      summary: Buscar y descargar documento
      description: |
        Busca un documento por criterios y lo descarga directamente en una sola operación.
        
        Primero realiza una búsqueda en la base de datos interna con los criterios proporcionados,
        luego recupera el documento desde Alfresco utilizando el ID encontrado.
      parameters: []

      requestBody:
        description: Criterios de búsqueda del documento
        content:
          application/json:
            schema:
              type: object
              properties:
                rut_cliente:
                  type: string
                  description: RUT del cliente (formato MongoDB)
                  example: "05393768-5"
                tipo_documento:
                  type: string
                  description: Tipo de documento (formato MongoDB)
                  example: "Contrato"
                nombre_documento:
                  type: string
                  description: Nombre del documento (formato MongoDB)
                fecha_carga:
                  type: string
                  description: Fecha de carga (formato MongoDB)
                razon_social:
                  type: string
                  description: Razón social (formato MongoDB)
                "tanner:rut-cliente":
                  type: string
                  description: RUT del cliente (formato Alfresco)
                  example: "05393768-5"
                "tanner:tipo-documento":
                  type: string
                  description: Tipo de documento (formato Alfresco)
                  example: "Contrato"
                "tanner:nombre-doc":
                  type: string
                  description: Nombre del documento (formato Alfresco)
                "tanner:fecha-carga":
                  type: string
                  description: Fecha de carga (formato Alfresco)
                "tanner:razon-social-cliente":
                  type: string
                  description: Razón social (formato Alfresco)
                "tanner:estado-vigencia":
                  type: string
                  description: Estado de vigencia del documento
                  example: "Vigente"
            examples:
              Busqueda por propiedades tanner:
                value:
                  "tanner:nombre-doc": "Contrato Servicios 2024"
                  "tanner:tipo-documento": "Contrato"
                  "tanner:razon-social-cliente": "Nombre empresa 1"
                  "tanner:rut-cliente": "05393768-5"
                  "tanner:estado-vigencia": "Vigente"
                  "tanner:fecha-carga": "2024-03-04T12:00:00.000Z"

              Busqueda por propiedades mongoDB:
                value: 
                  "tipo_documento": "Contrato"
                  "rut_cliente": "05393768-5"
                  "nombre_documento": "Contrato Servicios 2024"
                  "fecha_carga": "2024-03-04T12:00:00.000Z"
                  "razon_social": "Nombre empresa 1"
      responses:
        200:
          description: Contenido binario del documento
          content:
            application/pdf:
              schema:
                type: string
                format: binary
              examples:
                DocEncontrado:
                  value:
                    Documento.pdf
        400:
          description: Error de validación de entrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBase"
              examples:
                RUTFaltante:
                  value:
                    error: "Se requiere el campo rut_cliente o tanner:rut-cliente"
                    code: 400
        401:
          description: Problemas de autenticación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBase"
              examples:
                TicketFaltante:
                  value:
                    error: "Se requiere el header Authorization"
                    code: 401
        404:
          description: Documento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBase"
              examples:
                DocNoEncontrado:
                  value:
                    error: "No se encontraron documentos con los criterios proporcionados"
                    code: 404
        500:
          description: Error de procesamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBase"
              examples:
                ErrorDescarga:
                  value:
                    error: "Error al recuperar el archivo desde el repositorio"
                    code: 500
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: Autenticación básica con ticket de acceso
...
  schemas:
    ErrorBase:
      type: object
      description: Estructura estándar para respuestas de error
      properties:
        error:
          type: string
          description: Mensaje descriptivo del error
          example: "Error de validación"
        code:
          type: integer
          description: Código HTTP asociado
          example: 400